+++ src/nxt_conn_connect.c
@@ -155,7 +155,8 @@ nxt_conn_connect_test(nxt_task_t *task, void *obj, void *data)
     c->socket.error = err;
 
     nxt_log(task, nxt_socket_error_level(err), "connect(%d, %*s) failed %E",
-            c->socket.fd, c->remote->length, nxt_sockaddr_start(c->remote));
+            c->socket.fd, (size_t) c->remote->length,
+            nxt_sockaddr_start(c->remote));
 
     nxt_conn_connect_error(task, c, data);
 }
